--- BE/src/main/java/kr/codesquad/issuetracker/domain/dto/UserOfIssue.java
@@ -1,18 +1,0 @@
-package kr.codesquad.issuetracker.domain.dto;
-
-import kr.codesquad.issuetracker.domain.entity.User;
-import lombok.Getter;
-import lombok.ToString;
-
-@Getter
-@ToString
-public class UserOfIssue {
-
-  private String nickname;
-  private String profileImage;
-
-  public UserOfIssue(User user) {
-    this.nickname = user.getNickname();
-    this.profileImage = user.getProfileImage();
-  }
-}
