--- BE/src/main/java/kr/codesquad/issuetracker/domain/entity/IssueAssignee.java
@@ -1,31 +1,0 @@
-package kr.codesquad.issuetracker.domain.entity;
-
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-import javax.persistence.JoinColumn;
-import javax.persistence.ManyToOne;
-import lombok.AccessLevel;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.ToString;
-
-@Getter
-@ToString
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-@Entity
-public class IssueAssignee {
-
-  @Id
-  @GeneratedValue(strategy = GenerationType.IDENTITY)
-  private Long id;
-
-  @ManyToOne
-  @JoinColumn(name = "assignee_id")
-  private User assignee;
-
-  @ManyToOne
-  @JoinColumn(name = "issue_id")
-  private Issue issue;
-}
