--- BE/src/main/java/kr/codesquad/issuetracker/domain/entity/Milestone.java
@@ -1,39 +1,0 @@
-package kr.codesquad.issuetracker.domain.entity;
-
-import java.time.LocalDate;
-import java.util.List;
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-import javax.persistence.OneToMany;
-import lombok.AccessLevel;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.ToString;
-
-@Getter
-@ToString
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-@Entity
-public class Milestone {
-
-  @Id
-  @GeneratedValue(strategy = GenerationType.IDENTITY)
-  private Long id;
-
-  private boolean isOpened;
-  private String title;
-  private String description;
-  private LocalDate dueDate;
-
-  @OneToMany(mappedBy = "milestone")
-  private List<Issue> issues;
-
-  @Builder
-  private Milestone(Long id, String title) {
-    this.id = id;
-    this.title = title;
-  }
-}
