+++ BE/src/main/java/kr/codesquad/issuetracker/domain/label/LabelOfIssue.java
@@ -1,0 +1,17 @@
+package kr.codesquad.issuetracker.domain.label;
+
+import lombok.Getter;
+import lombok.ToString;
+
+@Getter
+@ToString
+public class LabelOfIssue {
+
+  private String title;
+  private String color;
+
+  public LabelOfIssue(Label label) {
+    this.title = label.getTitle();
+    this.color = label.getColor();
+  }
+}
