+++ BE/src/main/java/kr/codesquad/issuetracker/domain/user/UserOfIssue.java
@@ -1,0 +1,17 @@
+package kr.codesquad.issuetracker.domain.user;
+
+import lombok.Getter;
+import lombok.ToString;
+
+@Getter
+@ToString
+public class UserOfIssue {
+
+  private String nickname;
+  private String profileImage;
+
+  public UserOfIssue(User user) {
+    this.nickname = user.getNickname();
+    this.profileImage = user.getProfileImage();
+  }
+}
